#169 - Moved the code from the Platforms project to the Core project. Creating the platforms project was unnecessary.
