+++ src/Sidekick.Core/Natives/IKeybindEvents.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Threading.Tasks;
 
-namespace Sidekick.Platforms
+namespace Sidekick.Core.Natives
 {
     public interface IKeybindEvents
     {
