+++ src/Sidekick.Core/Natives/INativeBrowser.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Sidekick.Platforms
+namespace Sidekick.Core.Natives
 {
     public interface INativeBrowser
     {
