+++ src/Sidekick.Core/Natives/INativeProcess.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Sidekick.Platforms
+namespace Sidekick.Core.Natives
 {
     public interface INativeProcess
     {
