+++ src/Sidekick.Core/Natives/KeyboardCommandEnum.cs
@@ -1,4 +1,4 @@
-namespace Sidekick.Platforms
+namespace Sidekick.Core.Natives
 {
     public enum KeyboardCommandEnum
     {
