+++ src/Sidekick.Core/Natives/NativeClipboard.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace Sidekick.Platforms
+namespace Sidekick.Core.Natives
 {
     public class NativeClipboard : INativeClipboard
     {
