+++ src/Sidekick.Core/Sidekick.Core.csproj
@@ -13,6 +13,7 @@
     <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.1" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
     <PackageReference Include="System.Text.Json" Version="4.7.0" />
+    <PackageReference Include="TextCopy" Version="3.0.0" />
   </ItemGroup>
 
 </Project>
