+++ src/Sidekick.Platforms.Windows/Natives/NativeKeyboard.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using Sidekick.Core.Initialization;
 using Sidekick.Core.Loggers;
+using Sidekick.Core.Natives;
 using Sidekick.Core.Settings;
 using Sidekick.Platforms.Windows.Natives.Helpers;
 using WindowsHook;
