+++ src/Sidekick.Platforms.Windows/Natives/NativeProcess.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using Sidekick.Core.Initialization;
 using Sidekick.Core.Loggers;
+using Sidekick.Core.Natives;
 
 namespace Sidekick.Platforms.Windows.Natives
 {
