+++ src/Sidekick.Platforms.Windows/Sidekick.Platforms.Windows.csproj
@@ -15,7 +15,6 @@
   
   <ItemGroup>
     <ProjectReference Include="..\Sidekick.Core\Sidekick.Core.csproj" />
-    <ProjectReference Include="..\Sidekick.Platforms\Sidekick.Platforms.csproj" />
   </ItemGroup>
 
 </Project>
