--- src/Sidekick.Platforms/Sidekick.Platforms.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="TextCopy" Version="3.0.0" />
-  </ItemGroup>
-
-</Project>
