+++ src/Sidekick.UI/Sidekick.UI.csproj
@@ -7,7 +7,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Sidekick.Business\Sidekick.Business.csproj" />
     <ProjectReference Include="..\Sidekick.Core\Sidekick.Core.csproj" />
-    <ProjectReference Include="..\Sidekick.Platforms\Sidekick.Platforms.csproj" />
   </ItemGroup>
 
 </Project>
