+++ src/Sidekick/Windows/Settings/SettingsView.xaml.cs
@@ -4,8 +4,8 @@ using System.Windows;
 using System.Windows.Forms.Integration;
 using System.Windows.Input;
 using Sidekick.Business.Languages.UI;
+using Sidekick.Core.Natives;
 using Sidekick.Core.Settings;
-using Sidekick.Platforms;
 using Sidekick.UI.Settings;
 using Sidekick.UI.Views;
 using Sidekick.Windows.Settings.UserControls;
